pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentbridge"
version = "0.1.0"
description = "Universal AI Agent Interoperability Protocol"
authors = [
    {name = "TheKoma-X"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "websockets>=11.0.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "typing-extensions>=4.5.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.0.0"
]

[project.scripts]
agentbridge = "agentbridge.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentbridge*"]